move statvfs querying to linux mod
